[NO-TICKET] Fix deploy-demo script and convert it to TypeScript (#2077)
* Update renamed command name in deploy-demo.sh

* Create a TypeScript version of the deploy demo script

* Remove stdio config because it makes the code noisy and storybook doesn't conform

Storybook seems to write to stderr whenever it wants, which means we'd want a different stdio config for it. All this code makes it hard to read. And most of the time we're going to want to know what's happening and why it failed anyway. All that means I'm scrapping the verbose flag that turns off individual-command output

* Whoops, can't remove the stdio config entirely

* Huh, not sure why execSync wasn't working before

Maybe my system update fixed it

* Clean up and put stdio inherit config back because the output is nicer

* Remove old bash script and add some dependencies back

Co-authored-by: Forrest Baer <forrest@forrestbaer.com>