Upgrading RTL even further gets rid of focus-trap errors
I want to add another debugging note before I wrap this up. In the Tooltip implementation file, I was able to put some debug code in the `initialFocus` function which showed that the document was basically empty (an empty `<body>` and `<head>`) whenever it would throw an error. When it didn't throw an error, the document had everything that was expected. I'm glad it was fixed 😅