[WNMGDS-1903] Fix "cmsds-migrate" script failing in downstream projects (#2099)
* We need to include the scripts folder for `npm pack` to package it

I'm also removing the extra package file because it's not a package, and I think it's confusing things. I'm in the middle of troubleshooting failures in a downstream project that is still on node v12. Some findings:
- It mostly works if I change the file extensions to `.mjs`, though there are still some issues with incompatibility for their earlier version of node. The globby import is also a problem. I can make it work in the downstream project or ours but not both.
- If we can’t rely on our code running on all versions of node, we’re going to have to either 1) transpile down to something more basic or 2) write our script code in something more backwards compatible like CommonJS

* Switch to mjs files so node will treat them as such

Also import fs functions in a way that works with older versions of node

* String.replaceAll is not a function in node v12

* For globby to resolve to the right version downstream, it can't be a devDependency

It's unfortunate that we need to list these as dependencies now, but it will ensure that the migrate script works as expected in downstream projects

* Whoops, I accidentally removed a dependency