[NO-TICKET] Move test configs out of the `design-system-scripts` package (#2111)
* Move our test files out of the design-system-scripts package

* Moved test dependencies to root package.json

Note that versions of two babel things changed in the move because they already existed at the root. The result was minor version bumps, which should be fine.

* Hook up `yarn test:unit` to new config/command

I made the unit test config not wrapped in a function anymore and removed the module resolution override stuff. I think it might be running faster on average because of that change

* Remove cmsds cli command for unit tests

* Remove TODO comment that was outdated

* Hook up a11y test commands

* Remove cmsds cli command for a11y tests

* We need this module name mapper still so we don't have to build core first

* Default headless mode to true and fix unrecognized jest cli parameters

Jest CLI doesn't like extra "long options" that don't have values associated with them, so this is a hack to avoid having to put a wrapper around our jest command to parse options. We can revisit this later