Trying to get Playwright to work in docker with Storybook
The fact that the whole test runner runs inside docker I think is going to be a problem on Jenkins. It's going to be a complex docker setup there because there isn't really a concept of a host system. I may have better luck using Puppeteer where I've successfully set up Jenkins before such that the docker container only has a chrome instance that our test runner can connect to.