Update other colors that were changed
which we've determine to have only a very small impact.

We tracked down where they were used and concluded that `primary-dark` and `primary-darker` have the biggest impact, and it’s only on hover, focus, or active states for links and buttons—not on their resting states.

`primary-alt-*` and `secondary-alt-*` only affect the utility classes for color that go by the same names.

We also did a search in the repositories that are still using pre-8.0 versions of ds-healthcare-gov for any usage of those variables directly, and we found none.