[WNMGDS-1792] Dialog to native dialog (#2175)
* WIP: Converting Dialog to use NativeDialog

* Move analytics code into a separate hook

* Untested backdrop click code

* Get it to look like the old dialog

* Working on code that allows turning off ESC-exiting; don't think I'll keep it

The problem is that the `cancel` event is actually picking up on the event before the keydown, depending on whether the user has shifted focus from the default location after opening the dialog. I might be able to get it to work by adding some logic inside the `handleCancel` function, but I think it might be better to deprecate the `escapeExits` prop and avoid changing the default behavior of native dialogs in that way

* Don't allow `escapeExits` behavior

See previous commit message. I don't think it's a good idea to stray from the default native dialog behavior in this regard. And it simplifies our API and implementation not to have it

* Pass an id prop to the native dialog

* Allow ESC to exit idle timeout dialogs as well

This behavior was likely a holdover from the original component that was created in one of the apps, and my research with app teams would indicate that this is not an intentional or desired behavior

* Replace underlay* props with `backdropClickExits` prop

that passes through to the NativeDialog component. We can't easily apply styles to the backdrop/underlay anyway, and apps shouldn't have a need to change this. If they really want to change it, they can apply their own custom CSS rules

* Add a story for testing out scroll-preventing behavior (not implemented)

* Redo change from 69853b66

* Might have fixed the background scrolling issue

* Narrow the story content so we don't have to shrink the window to scroll

* We can have proper effect dependencies here

* We don't use loki anymore

* Fix the story controls so we can test `scrollDisabled`

* Mock `window.scrollTo` for tests

* Fix warning for missing exit function

* Fix errors in IdleTimeoutDialog.test.tsx

* Use `screen` in IdleTimeout tests

* Fix all but one of the failing IdleTimeout tests

* Fix last failing test for IdleTimeout

* Remove obsolete prop (caught by TypeScript)

* Remove react-aria-modal dependency

* Always disable scrolling the body and remove the option

* Put initial focus on the dialog body itself instead of the first interactive element

This matches the existing behavior more closely.

* Update snapshots after the initial-focus change

* Label the document div by the heading (which matches old behavior and is nice)

* Disable default focus halo for the role="document" div

* Fix bug in stories

* Update screenshots for dialog with focus change, and ignore the prevent-scroll story

* Build storybook fresh whenever we update snapshots

* Update the type to accept all valid `<dialog>` attrs

Co-authored-by: Forrest Baer <forrest@forrestbaer.com>