Removing alternate import path examples
I think that we want to encourage importing by name in most cases, especially if we want to be able to relocate files on our end. I think most teams are using modern build tools, and many will be familiar enough with these kinds of JavaScript performance concerns to make their own decisions.