None of this works. Saving what I have for future reference
I started gutting this local fork of the plugin and tried to insert the new code from my test script, but I was finding that docgen-typescript wasn't working in that folder. I attempted to get my test script (`docgen.js`) in the `packages/docs` root, but it fails because we use the root `tsconfig.json`, and that configuration is invalid relative to the `packages/docs` directory. We get the following error:

```
Error: TS18003: No inputs were found in config file '../../tsconfig.json'. Specified 'include' paths were '["packages/**/src/**/*"]' and 'exclude' paths were '[]'.
```