[WNMGDS-2140] Allow blog to have multiple articles (#2323)
* Fix TOC links in dev mode

Previously when clicking one of these links in dev mode, Gatsby would attempt to make a new page request instead of just scrolling to the right heading, but it results in a 404.

* (WIP) creating a framework for the blog pages

* Fixed a bug where the clientside router stripped trailing slashes

My full dev notes that lead up to this change:

I'm currently working on adding a blog section to our site, where individual blog article files in `/content/blog/` get rendered with their own template. However, I don't love the `memo()` HOC around the existing `InfoPage` template, so I'm getting distracted trying to figure out how to remove it.

We have an existing bug in development where if you click one of the in-page links from the table of contents (like `href="#react"`), it will 404. This only happens in development. I found that if I change the TOC links from `<a>` to the Gatsby `<Link>`, it will follow the links like  normal. However, the rendered anchor tag has an href with the page path in it, like `href="components/button/#react"`. In both cases, the Gatsby router normally wants to go fetch the page, even though it's an internal link. But when we make it a Gatsby `<Link>`, it at least knows where to find the page and doesn't throw a 404.

We don't actually want Gatsby to have to go talk to the router, though, and that's why we've had to add the `memo()` HOC. If I can somehow get Gatsby to recognize that it's an internal page link and that it doesn't have to do anything, we could get rid of this extra code for each page template. So far my internet searches have turned up with nothing. It's hard to believe nobody has run into this problem before. In-page links to id'ed elements on pages is pretty common, especially since there's a `gatsby-remark-autolink-headers` plugin. I even searched for this problem with `"gatsby-remark-autolink-headers"` in my query, but I got nothing relevant.

I've done some tests where I build it statically with and without the memoization, and in both cases the iframes on the page reload themselves when clicking internal links. I don't get it. What's making those iframes reload? The page itself is not reloading. It's not making a request to the server. It's something within the clientside application that is causing those iframe elements to re-render.

Oh!!! I know what is causing it. When we first load the page, the URL looks like this after hydration: `http://localhost:9000/components/alert?theme=core`. It would seem that the query (or `search`) string of `?theme=core` is the culprit, because none of the TOC links have that query string, but it's actually that `location.pathname` changes from `/components/alert` to `/components/alert/`. The following expression in the memo HOC evaluates to `false`.

```ts
nextLocation.pathname === prevLocation.pathname
```

For some reason even though we make the request for `...alert/` (with a trailing slash) and no redirect is made—we in fact redirect from no trailing slash to a trailing slash—the client removes the trailing slash! I'm trying to figure out what in Gatsby is removing that trailing slash. Gatsby is designed to have the trailing slash, because the files it builds puts everything in directories. It doesn't make sense why it would then be removing them clientside! It's not the theme query code, because even when I disable the call to `setQueryParam`, the trailing slash gets removed.

Just found out that if I added the trailing slash to the slug used in the `path` in the `createPage` function, the Gatsby router respects the trailing slash. I guess it makes sense that if we're the ones calling `createPage` explicitly, that that would be where we're in charge of saying whether it exists or not.

* Create a shared function for this feature

* The blog index page can list out titles and stuff

But right now we don't know what this page is going to look like and whether or not we will include any of the body text, so there's no use writing more of this page just yet

* Starting to shape how the blog pages look

* Fix link styles in headings and render the first article on blog index

* Add some lorem ipsum content for release 6 article

* Styled dedicated article pages with article listing on blog index

* Feedback widget and extra feedback link

* Make sure the section headings fit within measure--wide as well

Co-authored-by: Forrest Baer <forrest@forrestbaer.com>