Refining the unused healthcare button styles
For all but the alt solid, I just copied the `main` button styles to `alternate` because the alternate ones weren't fully fleshed out and are unused anyway, so there's no reason for them to be different. Because the alternate solid style (blue) is still in use on some applications, I left that alone.