[WNMGDS-2186] Creating `info` color category and refining `primary` and `secondary` (#2384)
* Add missing primary steps (each theme was missing different ones)

* Create an info color category and assign appropriate component vars to it

Note that I purposely changed the border color for filter chips in medicare to the secondary color to match the background that was secondary-lightest. I think this might be something nobody has noticed before because it's subtle.

* Got rid of green color variable

In core and healthcare, green shifted to secondary

* Refining the unused healthcare button styles

For all but the alt solid, I just copied the `main` button styles to `alternate` because the alternate ones weren't fully fleshed out and are unused anyway, so there's no reason for them to be different. Because the alternate solid style (blue) is still in use on some applications, I left that alone.

* Update variable used for utility example boxes in docs

This does result in a color shift, but they at least remain blue

* Whoops, tabs to spaces

* Update VRT refs for changed core buttons

* Update VRT refs for changed healthcare buttons

* Update the side bar color on the doc site

* Use lighter hover color for on dark solid alt buttons in core

* Create further consistency for on-dark core buttons

Now hover states always increase the contrast (lighter on dark, darker on light)

* Update VRT refs for core on-dark button changes

* Add theme colors to theme colors page

* Fix incorrect shade applied to `secondary-light`

Co-authored-by: Scott Weber <scott.weber@adhocteam.us>

---------

Co-authored-by: Scott Weber <scott.weber@adhocteam.us>