[WNMGDS-2191] Clearly differentiate between doc headings and example headings (#2395)
* Wrap iframe examples in `<section>` as well

We were doing this for embedded ones but not for iframes. This is hopefully an alternative to https://github.com/CMSgov/design-system/pull/2386. Both `<article>` and `<section>` work the same way as far as I can tell where the heading levels are supposed to be somewhat independent from the rest of the document. If this doesn't work, we can switch to `article` and see if that works. It's hard for me to tell if it helps or not. When opening the web rotor inside the `section`, it shows just the headings for the section, but it still shows all headings in the top level when you're not in the section.

* Use the word "example" in our example heading names to differentiate from doc headings

* Update VRT refs for the changed stories

* Reuse the heading constant

---------

Co-authored-by: Scott Weber <scott.weber@adhocteam.us>