When I import instead of export from, I get a much more verbose error
It makes it past the TS/babel compile step, and it fails at the webpack step, where it prints a long error explaining all the places it's looking for the `@cmsgov/design-system/web-components` module, all of which are incorrect. It doesn't seem to follow the `"./web-components":` entry in the core package.json file, and it complains that the "browser" config is blank.