Revert earlier color change after rethinking the role of the footer
I had originally changed this from `secondary` to `primary` when `secondary` colors moved into `info` because I thought a primary background would be more appropriate for the brand. However, this might have been an unnecessary color change for the component, especially considering that the footer may actually be used in an informational role, like an info alert. I'm not sure if that's true, but I'm proposing we don't actually change this value if we haven't figured out what our alternative background color palette looks like yet.