In order to appease webpack, we needed to use the full path instead of package.json exports path
However, if we `export * from` that full path to the web-components index file, TypeScript doesn't know where to get its type definitions, because the link
to the TS definitions for that entry point were established in that "exports" config in the package.json file, and using the full path circumvents it. Only
by using "import" and the full path do we circumvent both the package.json exports config so webpack works and circumvent type checking so TypeScript works.
 BUT WHEN WE DO THAT, THE COMPONENTS DON'T SHOW UP IN THE ds-healthcare-gov web-components BUNDLE!!!!