With a wee bit of code, we don't need `js-cookie`
This is more correct because we now have a `path` setting to make sure it always gets set at the top level. We can also use `max-age` by itself without `expires` because we don't need to support ancient IE versions. It's only a few more lines of code and gets rid of a dependency that is a lot bigger than that.