[WNMGDS-2245] Remove primary alt color (#2416)
* remove references in utility styles

* update storybook stories

* update themes

* update storybook styles

* Update theme files

* updating snapshots

* Update packages/design-system/src/components/utilities/margin.stories.tsx

Co-authored-by: Patrick Wolfert <patrick.wolfert@adhocteam.us>

* Update packages/design-system/src/components/utilities/padding.stories.tsx

Co-authored-by: Patrick Wolfert <patrick.wolfert@adhocteam.us>

* Remove last unrelated ref image change

* Update the refs with accurate colors

In order to get our tests to pick up on this change, I had to check these out from `main` and then run the tests again. Then the tests created the updated reference images and I was able to check them in

* Use `info` equivalent of old color so there are no story changes at all

* Undo the mask image change

Checked it out from main this time. Just trying to get it out of the PR diff

---------

Co-authored-by: Forrest Baer <forrest@forrestbaer.com>
Co-authored-by: Patrick Wolfert <patrick.wolfert@adhocteam.us>