[WNMGDS-1611] Move privacy settings modal to core (#2422)
* Move the privacy settings code as-is into core

* Update cookie domain logic after reading MDN docs

And also this didn't account for a missing subdomain in the `location.hostname` before. Now it does by starting from the end of the string.

* With a wee bit of code, we don't need `js-cookie`

This is more correct because we now have a `path` setting to make sure it always gets set at the top level. We can also use `max-age` by itself without `expires` because we don't need to support ancient IE versions. It's only a few more lines of code and gets rid of a dependency that is a lot bigger than that.

* Noticed that this dependency version didn't match core

* Update the cookie unit tests

* Move translated content into core

* Make it so the links to policy documentation can be changed

* Simplify the table styles by using what we have in the design system

The mobile view is where all the extra code is coming from, but I don't see why we wouldn't just use our existing stacked table that creates an accessible mobile experience

* Use the core component in healthcare

* Add visual regression test references

* Update unit tests and snapshots

* Fix console error by adding missing prop

* Update props used in tests

* Fix lack of `headers` prop in the stackable table story, which caused warnings

---------

Co-authored-by: Forrest Baer <forrest@forrestbaer.com>