Untested isChromatic logic that mirrors our VRT tests
But I think I might swap this out in favor of something a bit less magic. I think what I want to do instead is add a parameter to the healthcare and medicare stories for the theme and set it to the specific theme it's for. That way you're only ever going to see that story in that theme when you visit the storybook as well.