Configure `@percy/storybook` to take snapshots of stories in different themes!
This is one of the reasons we had to write our own Playwright-based solutions. Chromatic doesn't have any way of configuring it to take these additional snapshots with extra parameters. But `@percy/storybook` does it out of the box! It was so easy. I ran ` yarn percy storybook ./storybook-static`, and it listed all the snapshots it was taking, including the extra themed snapshots for each story

However, if we were to drop our Playright tests, we'd lose the forced-colors mode testing