[WNMGDS-2290] Switch from Downshift's old API to new and no longer clear Autocomplete input on blur (#2483)
* Start using the downshift hooks api in autocomplete

* Map the optional label heading to the menu like it was before

This `aria-labelledby` value replaces the Downshift default. I think it would be better to use both this heading and the form label, but in this first commit I want to restore what was there before

* Use both the heading and form label as labels

* Apply the additional autocomplete props to Downshift

* Try to improve code legibility

* Standardize ids a bit, kind of matching our BEM pattern of class names

* Fix TypeScript errors and map new API to old API

* Unit tests found behavioral change/regression for `items={undefined}`

* Fix all but 1 test. Remaining test found a regression

The default behavior of downshift seems to have changed so that when ESC is pressed, it doesn't clear the input. Previously we needed a custom state reducer to keep it from clearing the input

* Don't clear input on ESC anymore. It's not helpful

* Remove default behavior of clearing input text on blur

Discussed this with application teams at CMS, and everyone was in agreement that this is not helpful behavior and in fact is behavior that could be frustrating to users.

* ESC no longer clears text input, and we want it that way

* Clean up types

* Move some arcane code into its own file to improve readability

* Remove redundant snapshot

* Restore `onChange` functionality and patch holes in unit tests

* Provide a more accurate type def for `onChange`