Fix bug in `findElementsOfType` that caused plain option children to error
When testing out a story where the children were just `option` elements without groups, it would throw an error when one of the option's `children` was a string. The cases in the function are hopefully way clearer now, and TypeScript is much happier and is able to check that everything is what it should be at each step without having to type cast!