Go back to including the opt groups in the downshift items array
to get a smoother interaction with pointer devices. As a pointer hovers over the group headings, the hover styles stay on the last hovered item rather than the hover style disappearing when I hover over the group heading. The highlightedIndex doesn't get updated as the pointer passes over the group headings, because they don't exist as items in Downshift. It looks like it "sticks", which feels clunky. They both exhibit the same behavior when using the keyboard, which is to skip the group, so keyboard interaction is not a factor in this decision.