[WNMGDS-2317] Re-implement `Dropdown` as combobox using `Downshift` (#2498)
* Some basic functionality with lots of decisions yet to be made

* A little more progress and cleanup

* The `getItemProps` function handles setting `aria-selected` already

* Some basic interaction styling for the dropdown menu

* Add support for size prop

* This logic for parsing optgroups and options works but might not be robust enough

* Rendering opt groups mostly works, but you can still keyboard highlight the groups

I think maybe in order to get this to work correctly, I need to exclude the groups from the `items` that I pass the downshift hook

* Mostly working except that the highlightedIndex doesn't get set to selected item index after selection

* Add some comments

* Pass any extra props on options and optgroups to the list items

and move a bunch of functions into a utils module

* Move non-prod prop validation code into separate module

* Still unsure why `highlightedIndex` resets itself after selection

* Some fancier styling

* Cleanup from pairing session

* Basic styles to differentiate groups

* Fix bug in `findElementsOfType` that caused plain option children to error

When testing out a story where the children were just `option` elements without groups, it would throw an error when one of the option's `children` was a string. The cases in the function are hopefully way clearer now, and TypeScript is much happier and is able to check that everything is what it should be at each step without having to type cast!

* Saving the `highlightedIndex` after selection works!

Had to memoize the options since we generate them each time when we parse children

* Go back to including the opt groups in the downshift items array

to get a smoother interaction with pointer devices. As a pointer hovers over the group headings, the hover styles stay on the last hovered item rather than the hover style disappearing when I hover over the group heading. The highlightedIndex doesn't get updated as the pointer passes over the group headings, because they don't exist as items in Downshift. It looks like it "sticks", which feels clunky. They both exhibit the same behavior when using the keyboard, which is to skip the group, so keyboard interaction is not a factor in this decision.

* Clean up types and fix ts error

* Fixing up tests and snapshots and patching holes that they find

* Fix remaining unit tests

* More unit tests about children and optgroups

* Add story for array-based option groups

* Update visual regression tests, add interaction tests, and update some snapshots

There are some snapshot updates that I have not included in this commit because they actually found regressions that I need to fix with the medicare styles

* Fix medicare dropdown icon regression

* Fix regressions and update snapshots

* Revert Autocomplete changes in favor of changes on `main`

* Fix references to `select` in props docs

* Remove this type annotation in doc site version switcher because it's wrong now

* Fix storybook links

---------

Co-authored-by: Forrest Baer <forrest@forrestbaer.com>