[WNMGDS-2373] Update ThirdPartyExternalLink to use anchor instead of button (#2525)
* Use link after talking to a11y experts, and update styles

Our argument for using a link is that even though it doesn't perform the navigation directly (normal a11y rule for using a button instead of a link), the intent of the element is to perform navigation, which makes it a link. This also allows desktop users to hover over the link and see the address, and it allows some other standard browser behaviors. It is also way easier to style to look like a link, because it is a link. We realized we were going to have to fix the styles anyway because its primary use case is to be within blocks of text.

* Update unit tests and snapshots

* Update VRTs