Fix example width overflow with a rewrite
Notes:

The root issue is that even though we’re scaling the wrapper around the iframe on wider simulated breakpoints, transform scale doesn’t actually mean the element becomes that transformed width in the flow of the document; it just looks smaller. Therefore it ends up widening the page by a lot when we choose the lg or xl example sizes, and there’s all this dead space to the right of it

Figuring out what the natural width should be is pretty tricky. I think what we’ll want to do is calculate the natural width of the breakpoint buttons on top before rendering the iframe, and then we have a fixed width to make that wrapping container so it doesn’t mess with all the stuff on the page. Even that might pose a problem when we start to resize the window

Ended up taking the scaled example outside of the normal document flow by making it absolute