[WNMGDS-2384] Revert heading spacing changes (#2544)
* Revert "[WNMGDS-2377] Remove heading bottom margins (#2534)"

This reverts commit 4a29bbfa8362fb1fca535bd1ae012c711bbd5239.

* Remove headings from list of spacing changes

But we still need to update the example of margin collapse

* Remove `:first-child` rules on base heading and paragraph elements

We've gotten feedback that they interfere in places they shouldn't. Those rules assumed they would always be used in content areas, but really that should be left up to the `ds-content` class or individual typography classes

* Update spacing migration diagrams

* Fix some regressions

* Use content class in all-button stories

Gets rid of extra margin at the top and makes the headings more legible

* Fix inconsistency in treatment of `sm` text body class

In the VRTs, I noticed the sm story was the only one with extra margin at the top. I think it makes more sense for the behavior of these body typography classes to be the same.

* Top margin is unnecessary in this story

* Fix typo in headings doc page

* Remove duplicate story that isn't referenced in docs

The real one is located in `packages/design-system/src/styles/utilities/font-size.stories.tsx`

* Name this story better

* Update VRT refs

I've looked at all of these, and they're expected

* Make the typography class messaging more meaningful