Re-enable stories on help drawer docs page
Now that the docs page isn't just a tab of a story preview, the help drawer doesn't really get in the way of the docs. The toggle button is there on the page, and you opt-in to pressing it. Therefore it doesn't get in the way of reading the docs. Also change the name of the default story for consistency