[WNMGDS-2402] More informative storybook doc pages (#2562)
* Add support for documenting stories with subcomponents

Before this if we had moved all props docs to storybook, some of the sub-component documentation would be lost

* Restrict theme styles to stories so doc heading styles don't get reset

There was a bug where the headgings on the storybook docs pages were getting hit by the reset styles in non-core themes because the `[data-theme=medicare] RULE` selectors had a higher specificity than the base Storybook classes

* Add basic doc comments for every component, which will show up in storybook

* Update ref image names and stories

In a previous commit, I tried to make the story folders more consistent by putting spaces between words. This finishes the job and fixes the tests

* Forgot to add a space in this one

* Rename more files

* Meant to rename this too but mistyped

* Revert "Meant to rename this too but mistyped"

This reverts commit 8ce82b5acb7a37f3b2fb9cbea4cb228fb2fd2f0a.

* Revert "Rename more files"

This reverts commit 6603eb3271b08fd78b6b4395c780098c9cc46515.

* Revert "Forgot to add a space in this one"

This reverts commit 659fec2be5f1b0ad53d8b2f31193eb4d006c7004.

* Revert "Update ref image names and stories"

This reverts commit c7db2f52c699cf7cd3309b37e1a9ab14f50df011.

* Fix steplist file names

* Re-enable stories on help drawer docs page

Now that the docs page isn't just a tab of a story preview, the help drawer doesn't really get in the way of the docs. The toggle button is there on the page, and you opt-in to pressing it. Therefore it doesn't get in the way of reading the docs. Also change the name of the default story for consistency

* Prevent privacy settings dialog from overlapping the docs page