Yet another attempt to import stories (statically transforming files first)
The idea here is that we'd replace imports in story files with mocks of the storybook api so we could render stories with the components and the args without all the overhead of storybook itself. Unfortunately I'm getting errors like this:

```
Error: Unable to resolve a valid export for "Story"! Please open an issue at https://astro.build/issues!
    at Module.generateHydrateScript (/Users/patrickwolfert/Projects/design-system/node_modules/astro/dist/runtime/server/hydration.js:79:11)
```

which I think means Astro has successfully made the story a clientside component but has been unsuccessful at bundling the right client code to hdyrate it again