Add item to PR checklist for adding a milestone
We don't always need to add a milestone, but this is a reminder to do so if it's applicable. A scenario where we wouldn't want to is if it's a change that doesn't need to show up in the release notes. In that case we wouldn't bother adding labels either.