[NO-TICKET] Tooling updates related to release process (#2584)
* Remove Forrest from bump reviews so he doesn't get spammed

* Add item to PR checklist for adding a milestone

We don't always need to add a milestone, but this is a reminder to do so if it's applicable. A scenario where we wouldn't want to is if it's a change that doesn't need to show up in the release notes. In that case we wouldn't bother adding labels either.

* Allow PRs to have arbitrary extra labels in release notes script

To test this, change line 47 to `const milestoneJSON = execSync('gh api repos/CMSgov/design-system/milestones?state=closed').toString();` and select the `7.0-beta.1` milestone. One of those PRs has an extra, unrelated label

* Optimize output for copy/pasting into slack, and don't lose extra ticket numbers

Same method of testing as previous commit (and same milestone). You should now see the following pair in the jira tickets output:

```
- [WNMGDS-2222](https://jira.cms.gov/browse/WNMGDS-2222) - Build react, preact, and web-components versions of each design system
- [WNMGDS-2224](https://jira.cms.gov/browse/WNMGDS-2224) - Build react, preact, and web-components versions of each design system
```

They came from the same PR