Switch gears and work on aria-labels. Put them on fieldsets instead
Dan tested on multiple screen-reader/browser combos and came to a preliminary conclusion that putting the aria-describedby attribute on the fieldsets performed better than putting them on the multiple inputs.

I'm putting most of this logic in the `useFormLabel` hook, but I'd still like to get rid of that abstraction soon.