Separate the checked children from the grid by bringing back the old DOM structure partially
Unfortuantely there's no robust way of getting the checked children elements to display outside the checkbox grid without putting it in a separate element entirely, so I'm having to bring some of the old DOM structure back.

Also changed the stories a bit because placing the `ds-c-choice__checkedChild` class directly on the alerts had a weird effect in other themes. It's not really intended to be combined with other styles anyway. It should just be a container element. Keeping that class off the individual choice story entirely, though, to show what a plain checked child looks like, since they tend to do that in medicare apps (bare alerts under choices).