Allow medicare label top spacing and line height to be relative to its smaller font size
Let me explain what lead to this decision.

The reason the top margin changed is that we set the top-margin of all labels to be `1.5em`, which is relative to the label’s font size, but medicare does some overriding of font sizes. Previously the medicare override applied the smaller font size to a span inside the label, not the label itself. I’ve simplified and removed the span, because it was only necessary to have a separate span for the label text when it had hint text and error message as siblings inside the label. However, that _small_ font size multiplied by `1.5em` is `21px` of margin instead of the `24px` that it would be if we were multiplying with the _normal_ font size. (Those pixels are relative to my browser settings because our fonts are now defined in rems). Medicare labels use the Rubik font at small (0.875rem (14px in my browser)) size, and they are rendered in ALL CAPS.

These were our options:
1. Increase relative margins for medicare labels to end up calculating out to the old value (24px)
2. Make medicare labels use a larger font size than they do now
3. Keep the medicare label font size how it is, and allow the top margins to become smaller so they stay `1.5em` relative to the actual font size (which would make them now 21px in my browser)

The reason we'd go for #1 would be to avoid any visual changes. However, it's very difficult to avoid _all_ visual changes with this change anyway, especially if you now add on top of it that the line-height of the label changed. The line-height of medicare labels used to implicitly use the line-height of its parent with a normal font size even though neither element was setting it, but now that I’ve taken out the span, the line-height is calculated from the smaller text size. Not sure how it was larger before, but now we have to consider whether we want it to be relative to its actual font size (`0.875rem`) or hack it to maintain its old value that was relative to `1rem`.

We decided to go with option 3 for the top margin and to allow the line-height to naturally change with the smaller font size. This is most consistent with other themes.