Fix dropdown item roles being clobbered by a wrong assumption
I assumed that if we passed `undefined` for the value of that role property to the `getItemProps`, the default value would come through, but it wasn't.

Axe DevTools flags the missing role and throws 17 errors when you have a dropdown open with these missing roles. When this is fixed, those errors go away.