Remove the gap for unstyled lists to avoid breaking other people's stuff
While we do want styled lists in text content to be uniform across products, we don't need to unnecessarily target other components that use lists and apply extra gap, padding, and margin. Maybe I'll also remove the margin in another commit. We were already removing the padding. This is for situations like the vertical nav where we do semantically want to use a list, but we want to apply our own styles. We've had reports of teams needing to apply different `flex-direction` values and remove `gap`. We don't really want to mess with these styles if we don't have to. While we knew we _would_ come into conflict, we hoped it'd be easy to for teams to overcome and move past. That was possibly a mistake.