[WNMGDS-2427] Filter-dialog prototype for doc site theme and version switching (#2599)
* We never did actually export `NativeDialog` even though it looked like it

Remove the export line from our index so it's not confusing. The export never worked. We haven't released this component to the public.

* Basic, untested FilterDialog commponent added to docs package

* WIP: Started with a dumb component and logical component, but...

The problem with that is that the versions list is dynamic; it's based on the currently selected theme. That means I need logic inside the component that renders the dialog to dynamically change the options in the version dropdown based on the theme

* Moved it all into one component, and it makes sense

* Hook it up and start working out kinks in version/theme switcher

Problem right now is that if you previously had one theme selected and you select a different one, it will think you changed versions too because the version numbers don't match. While I could convert all versions to the core version equivalent to compare, what if it can't find the corresponding core version and has to fall back? Then the comparison would be wrong. But that doesn't happen, because there are no versions that don't have core equivalents. So maybe that'll actually work

* Fixed `useTheme` causing a second render for every component that uses it

It would always start with "core" and then switch to the actual current value. That second render was causing problems in the theme switcher

* Reorganize modules and create a basic ThemeVersionSection for the sidenav

* Use inverse colors

* Fix bug in my useTheme fix

* Some work on styling

* Styling for the triggering button

* Add arrow icon to the "Change settings" button

* Use theme `displayName` in sidebar

* Remove old theme and version switcher components

* Turns out we don't need the onExit function

* Add some basic unit tests

* Try to fix lost selection message with custom getA11ySelectionMessage fn

Unfortunately it does not work because of the debouncing done inside Downshift. [Here's my explanation](https://github.com/downshift-js/downshift/issues/1244):

> For what it's worth, I have just come across an issue with the existing a11y-message implementation and specifically [how updateA11yStatus is debounced](https://github.com/downshift-js/downshift/blob/master/src/hooks/utils.js#L85). In my application, I have one dropdown on a page that modifies the available options for another dropdown further down the page. (This isn't ideal from an accessibility standpoint, but [the WCAG docs](https://www.w3.org/WAI/WCAG21/Understanding/on-input) do say we can do that if we warn users that it's going to happen, and it will take a while before we can design out all the instances of it across multiple applications.) I wanted to write a custom `getA11ySelectionMessage` function that would determine when it's one of those incidentally changed dropdowns rather than the one the user is interacting with so that only the desired `${itemToString(selectedItem)} has been selected.` gets printed to the a11y-message div. However, due to the debouncing, none of the other `getA11ySelectionMessage` functions even get called except for the last one, which is the one I don't want to announce. If Downshift called those `getA11yMessage` functions instead of passing them to `updateA11yStatus` and then ignored them if they returned, say, `undefined`, this plan would work. But right now the agency from each individual dropdown is taken from it to determine whether its status message should be announced. It's getting lost in the debounce.

* Get rid of new function that didn't work

See previous commit

* Pivot to using two dialogs

This should be a much more straightforward experience and not create accessibility problems with the theme dropdown dynamically changing the options of the version dropdown.

* Move focus back to the triggering button when closing a filter dialog

Need to figure out a clean way of making this the easy thing to do for teams

* Add a close button

* Add the border radius used for dropdowns

* Allow the close icon label to be announced

Forgot that it is off by default

* Show which design system theme we're on in the page (tab) title

* Fix HealthCare.gov theme text wrapping

Don't constrain the theme/version section more than we need to

* Update snapshot

* Better screen reader text for the theme and version buttons

* Move the theme and version section out of side nav on mobile

* Fine tune the spacing after making the hit box large enough

* Fix mobile spacing on page header

* Fixes according to feedback

* Don't let multiple dialogs be open at once

* Update this prop's documentation to make it more clear what they need to do

* Update snapshot